BaseAttachment's save method add isFile
変更がないのにファイルが消えてしまったから

+++ rise_info/rise_info/baseModels.py
@@ -79,7 +79,9 @@ class BaseAttachment(models.Model):
             self.file = upload_file
             if "force_insert" in kwargs:
                 kwargs.pop("force_insert")
-        self.file_delete()
+        else:
+            if not os.path.isfile(f'f"uploads/info/{str(self.id)}/{str(self.info)}/abc.jpg"'):
+                self.file_delete()
         super().save(*args, **kwargs)
 
     def delete(self, *args, **kwargs):
